rng refactor
